Store appointment data in appointment object
